refactor: 코드 포맷팅 및 프로젝트 구조 변경
- 의미 없는 공백 제거
- 프로젝트명 변경에 따른 패키지 구조 변경
- Callback 클래스명 오타 수정
