+++ src/main/java/org/example/bridgeserver/BridgeServerApplication.java
@@ -1,13 +1,12 @@
-package org.example.queueingserver;
+package org.example.bridgeserver;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class QueueingServerApplication {
+public class BridgeServerApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(QueueingServerApplication.class, args);
+        SpringApplication.run(BridgeServerApplication.class, args);
     }
-
 }
