+++ src/main/java/org/example/bridgeserver/listener/ShutdownEventListener.java
@@ -1,4 +1,4 @@
-package org.example.queueingserver.listener;
+package org.example.bridgeserver.listener;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
