+++ src/main/resources/application.properties
@@ -1,2 +1,2 @@
 spring.profiles.active=dev
-spring.application.name=RabbitMq-server
+spring.application.name=bridge-server
