+++ src/test/java/org/example/bridgeserver/BridgeServerApplicationTests.java
@@ -1,10 +1,10 @@
-package org.example.queueingserver;
+package org.example.bridgeserver;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class QueueingServerApplicationTests {
+class BridgeServerApplicationTests {
 
     @Test
     void contextLoads() {
